requirements.txt
Flask==2.2.5
psycopg2-binary==2.9.9
Werkzeug==2.2.3
python-dotenv==1.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
email-validator==1.3.1
razorpay==1.4.3
Flask-Mail==0.9.1
Flask-Session==0.4.0
